V1 Collision Editor functionality - You can select a channel and see it on the map - Meshbuilding is fast! - Got alpha working so now the backgrounds are pink again - Added some new bitmath stuff to TileSpriteInfo - Fixed a bug with saving map data ignoring the collision flags
Todo: It currently only shows the selected collision channel, not if it contains one of the channels. For example if you have 1 and 2 selected, only tles with BOTH flags 1 and 2 will be shown. Emily requested that there be another mode that if the tile has flags 1 and 2 but we have 1 selected the tile should still be shown. Like a strict vs loose filter of tiles.