Finished up v1 of my lua library. - Got AABB2, Vec2, Disc all classes - Added Berry namespace with functions that use these new classes for ease - Put the code inside the engine in a string so it is built into the engine
Fixed a bug where include order mattered. Now we put all content of an include at the top of the file instead of the bottom.
Fixed a bug where we couldn't create a class because metatable wasn't defined. Didn't realize that that wasnt part of the core lua so moved up where we add lua libraries and that now works

Added my test project for the lua library